refactor(shared): extract task enum constants into a decorator-free module
